requirements.txt
torch
pillow
openai-clip

